+++ package.json
@@ -50,6 +50,7 @@
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-react": "^7.34.1",
-    "prettier": "^3.2.5"
+    "prettier": "^3.2.5",
+    "typescript-plugin-css-modules": "^5.1.0"
   }
 }
