+++ src/app/signup/ui/SignupFrom.module.css
@@ -1,0 +1,8 @@
+.SignupForm{
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+
